backend/requirements.txt
# Web framework
fastapi
uvicorn[standard]
python-multipart
slowapi

# Database
sqlalchemy
psycopg2-binary
pgvector
alembic

# Document parsing & chunking
unstructured-client

# Embeddings
voyageai

# AI
anthropic

# Logging
structlog

# Config
python-dotenv

# Auth
PyJWT
